Solve issue #2: icons doesn't load
Fontaswesome icons are not loading due to the fact that font-awesome.css
uses paths for loading font files that are not valid in the context of
the content script. Solve by using a workaround that consists in
adding a style node in the head node of the DOM and making the
fontawesome library directory accesible for the content script (in the
manifest file).